Center method and insertion of randomly generated points now with margin

+++ src/propra22/q8493367/entities/ManhattanMetric.java
@@ -11,6 +11,7 @@ public class ManhattanMetric implements IMetric {
     
 	@Override
 	public double distance(int x1, int y1, int x2, int y2) {
-		return Math.abs(x2 - x1) + Math.abs(y2 - y1);
+		// Cast to double to avoid overflow
+		return Math.abs((double)x2 - (double)x1) + Math.abs((double)y2 - (double)y1);
 	}
 }
